Add constructor methods to LanguageGraph
