--- maltoolbox/language/compiler/requirements.txt
@@ -1,2 +1,0 @@
-antlr4-tools
-antlr4-python3-runtime
