+++ pyproject.toml
@@ -12,7 +12,9 @@ readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
   "py2neo>=2021.2.3",
-  "python-jsonschema-objects>=0.4.1"
+  "python-jsonschema-objects>=0.4.1",
+  "antlr4-tools",
+  "antlr4-python3-runtime"
 ]
 license = {text = "Apache Software License"}
 keywords = ["mal"]
