+++ requirements.txt
@@ -3,3 +3,7 @@ python-jsonschema-objects
 
 #maltoolbox.ingestors:
 py2neo
+
+#compiler
+antlr4-tools
+antlr4-python3-runtime
